Fix unused variable warnings; replace int_fast* types
darcs-hash:20050503175147-10d61-4e59f1513616fa0e2e4f1355721c12e0b70b057b.gz

+++ gp32/gpsound.h
@@ -20,7 +20,7 @@
 
 #include "../types.h"
 
-extern volatile uint_fast8_t playing_frame;
+extern volatile unsigned int playing_frame;
 
 void gpsound_init(uint32_t pclk, uint32_t *rate);
 uint16_t **gpsound_buffers(int size);
