+++ m6809.h
@@ -29,13 +29,13 @@ typedef union {
 	} byte_values;
 } accumulator_t;
 
-extern uint_fast8_t nmi, firq, irq;
+extern unsigned int nmi, firq, irq;
 
 void m6809_init(void);
 void m6809_reset(void);
 void m6809_cycle(Cycle until);
 void m6809_get_registers(uint8_t *regs);
 void m6809_set_registers(uint8_t *regs);
-void m6809_jump(uint_fast16_t pc);
+void m6809_jump(uint_least16_t pc);
 
 #endif  /* __M6809_H__ */
