+++ tape.h
@@ -13,6 +13,6 @@ void tape_reset(void);
 int tape_attach(char *filename);
 void tape_detach(void);
 int tape_autorun(char *filename);
-uint_fast8_t tape_read_bit(void);
+unsigned int tape_read_bit(void);
 
 #endif  /* __TAPE_H__ */
